Fixup VersionChecker usage.
* Ember is "singleton" (aka we are _always_ using the top level version,
and there can't me multiple nested ones), so swap to using
`VersionChecker(this)` to grab the Ember version.
* Using the argument to `included` will not work for top-level addons
(ones included in the project as direct deps or devDeps) because it has
no `.root` property which is what `ember-cli-version-checker` uses. So
change to `this.parent` when building up our "parent checker".