[temp bugfix] Allows compatibility helpers to work when linking
Currently, ember-cli-version-checker does not work when linking
packages due to the way it treats addons, roots, and projects. There are
some open PRs on how to fix this, but in the meantime this is a
temporary patch which should be backwards compatible, and force the
version checker to always find the Ember version at the actual root of
the project, regardless of linking.